requirements.txt
numpy
influxdb-client
python-dotenv